refactor: 섬 관련 기능 game-island로 분리
